requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1

# Google Gemini
google-generativeai==0.8.0

# Vector DB
chromadb==0.5.5

# Embeddings & Re-ranker
sentence-transformers==3.0.0

# BM25
bm25s==0.2.0

# Document Loading
pypdf==4.3.0
docx2txt==0.8
beautifulsoup4==4.12.3
httpx==0.27.0

# Tokenization
tiktoken==0.7.0

# Retry Logic
tenacity==8.5.0

# Observability
langfuse==2.40.0

# Evaluation & Data
pandas==2.2.0
numpy==1.26.4

# CORS
